requirements.txt
casadi
numpy
pandas
matplotlib
pyyaml